Buffer only for v2
