+++ test/unit/entities/DutchQuote.test.ts
@@ -23,8 +23,6 @@ import {
   FLAT_PORTION,
   PORTION_BIPS,
   PORTION_RECIPIENT,
-  SWAPPER,
-  TOKEN_IN,
 } from '../../constants';
 import {
   CLASSIC_QUOTE_EXACT_IN_LARGE,
@@ -116,142 +114,6 @@ describe('DutchQuote', () => {
 
       expect(amountInSlippageAdjusted.gte(amountInGasAdjusted)).toBeTruthy();
       expect(amountOutSlippageAdjusted.eq(amountOutGasAdjusted)).toBeTruthy();
-    });
-
-    it('buffer is subtracted from the output amounts for EXACT_INPUT', async () => {
-      const amount = BigNumber.from(AMOUNT);
-      const input = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const output = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
-        input,
-        output,
-        TradeType.EXACT_INPUT,
-        10
-      );
-      expect(bufferedInput.startAmount).toEqual(amount);
-      expect(bufferedInput.endAmount).toEqual(amount);
-      expect(bufferedOutput.startAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS - 10)
-          .div(BPS)
-      );
-      expect(bufferedOutput.endAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS - 10)
-          .div(BPS)
-      );
-    });
-
-    it('neg buffer is added to the output amounts for EXACT_INPUT', async () => {
-      const amount = BigNumber.from(AMOUNT);
-      const input = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const output = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
-        input,
-        output,
-        TradeType.EXACT_INPUT,
-        -10
-      );
-      expect(bufferedInput.startAmount).toEqual(amount);
-      expect(bufferedInput.endAmount).toEqual(amount);
-      expect(bufferedOutput.startAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS + 10)
-          .div(BPS)
-      );
-      expect(bufferedOutput.endAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS + 10)
-          .div(BPS)
-      );
-    });
-
-    it('buffer is added to the input amounts for EXACT_OUTPUT', async () => {
-      const amount = BigNumber.from(AMOUNT);
-      const input = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const output = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
-        input,
-        output,
-        TradeType.EXACT_OUTPUT,
-        10
-      );
-      expect(bufferedOutput.startAmount).toEqual(amount);
-      expect(bufferedOutput.endAmount).toEqual(amount);
-      expect(bufferedInput.startAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS + 10)
-          .div(BPS)
-      );
-      expect(bufferedInput.endAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS + 10)
-          .div(BPS)
-      );
-    });
-
-    it('neg buffer is subtracted from the input amounts for EXACT_OUTPUT', async () => {
-      const amount = BigNumber.from(AMOUNT);
-      const input = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const output = {
-        token: TOKEN_IN,
-        startAmount: amount,
-        endAmount: amount,
-        recipient: SWAPPER,
-      };
-      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
-        input,
-        output,
-        TradeType.EXACT_OUTPUT,
-        -10
-      );
-      expect(bufferedOutput.startAmount).toEqual(amount);
-      expect(bufferedOutput.endAmount).toEqual(amount);
-      expect(bufferedInput.startAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS - 10)
-          .div(BPS)
-      );
-      expect(bufferedInput.endAmount).toEqual(
-        BigNumber.from(amount)
-          .mul(BPS - 10)
-          .div(BPS)
-      );
     });
 
     it.each([
