+++ test/unit/entities/DutchV2Quote.test.ts
@@ -1,10 +1,11 @@
 import Logger from 'bunyan';
 
 import { it } from '@jest/globals';
+import { TradeType } from '@uniswap/sdk-core';
 import { BigNumber } from 'ethers';
 import { BPS } from '../../../lib/constants';
-import { DEFAULT_LABS_COSIGNER } from '../../../lib/entities';
-import { AMOUNT, ETH_IN, TOKEN_IN } from '../../constants';
+import { DEFAULT_LABS_COSIGNER, DutchQuote } from '../../../lib/entities';
+import { AMOUNT, ETH_IN, SWAPPER, TOKEN_IN } from '../../constants';
 import { createDutchV2QuoteWithRequestOverrides } from '../../utils/fixtures';
 
 describe('DutchV2Quote', () => {
@@ -66,6 +67,142 @@ describe('DutchV2Quote', () => {
       const hash = order.hash();
       expect(hash).toBeDefined();
       expect(hash.length).toEqual(66);
+    });
+
+    it('buffer is subtracted from the output amounts for EXACT_INPUT', async () => {
+      const amount = BigNumber.from(AMOUNT);
+      const input = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const output = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
+        input,
+        output,
+        TradeType.EXACT_INPUT,
+        10
+      );
+      expect(bufferedInput.startAmount).toEqual(amount);
+      expect(bufferedInput.endAmount).toEqual(amount);
+      expect(bufferedOutput.startAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS - 10)
+          .div(BPS)
+      );
+      expect(bufferedOutput.endAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS - 10)
+          .div(BPS)
+      );
+    });
+
+    it('neg buffer is added to the output amounts for EXACT_INPUT', async () => {
+      const amount = BigNumber.from(AMOUNT);
+      const input = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const output = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
+        input,
+        output,
+        TradeType.EXACT_INPUT,
+        -10
+      );
+      expect(bufferedInput.startAmount).toEqual(amount);
+      expect(bufferedInput.endAmount).toEqual(amount);
+      expect(bufferedOutput.startAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS + 10)
+          .div(BPS)
+      );
+      expect(bufferedOutput.endAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS + 10)
+          .div(BPS)
+      );
+    });
+
+    it('buffer is added to the input amounts for EXACT_OUTPUT', async () => {
+      const amount = BigNumber.from(AMOUNT);
+      const input = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const output = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
+        input,
+        output,
+        TradeType.EXACT_OUTPUT,
+        10
+      );
+      expect(bufferedOutput.startAmount).toEqual(amount);
+      expect(bufferedOutput.endAmount).toEqual(amount);
+      expect(bufferedInput.startAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS + 10)
+          .div(BPS)
+      );
+      expect(bufferedInput.endAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS + 10)
+          .div(BPS)
+      );
+    });
+
+    it('neg buffer is subtracted from the input amounts for EXACT_OUTPUT', async () => {
+      const amount = BigNumber.from(AMOUNT);
+      const input = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const output = {
+        token: TOKEN_IN,
+        startAmount: amount,
+        endAmount: amount,
+        recipient: SWAPPER,
+      };
+      const { input: bufferedInput, output: bufferedOutput } = DutchQuote.applyBufferToInputOutput(
+        input,
+        output,
+        TradeType.EXACT_OUTPUT,
+        -10
+      );
+      expect(bufferedOutput.startAmount).toEqual(amount);
+      expect(bufferedOutput.endAmount).toEqual(amount);
+      expect(bufferedInput.startAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS - 10)
+          .div(BPS)
+      );
+      expect(bufferedInput.endAmount).toEqual(
+        BigNumber.from(amount)
+          .mul(BPS - 10)
+          .div(BPS)
+      );
     });
   });
 });
